# specs/001-ai-client-documentation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Warba Bank — AI-Powered Client Documentation API
  version: 1.0.0
  description: |
    Backend contract for Track 1 of the Warba Bank Corporate Banking AI Challenge.

    Governed by Constitution v1.0.0. Three contract-level invariants are worth stating
    before the endpoints, because they are the reason several endpoints look the way
    they do:

    1. **No auto-approval path exists.** `POST /documents/{id}/approve` is the only
       transition into `APPROVED`, it requires an authenticated RM, and it requires the
       caller to echo the `content_hash` of the exact version being approved. There is no
       endpoint, parameter, or scheduled process that approves a document any other way
       (Principle III, FR-026/FR-027).
    2. **Failure produces no document.** Generation, screening, and retrieval failures
       return a 4xx/5xx error, never a partial draft (Principle I, FR-037).
    3. **Audit is read-only over HTTP.** No endpoint mutates or deletes an audit event.
       The database role backing this API does not hold UPDATE or DELETE on the audit
       table (Principle VIII, FR-032).

    All data served by this API is synthetic (Principle VII).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                required: [access_token, token_type, user]
                properties:
                  access_token: { type: string }
                  token_type: { type: string, enum: [bearer] }
                  expires_in: { type: integer, description: Seconds }
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current authenticated user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ─────────────────────────── Clients ────────────────────────────
  /clients:
    get:
      tags: [Clients]
      summary: List clients in the caller's portfolio
      description: |
        An `RM` receives only clients where they are the owning RM. Scoping is applied by a
        shared authorisation dependency, not per-endpoint (FR-001, R11).
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/ClientSummary' }
                  total: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /clients/{client_id}/context:
    get:
      tags: [Clients]
      summary: Preview the context that would be used for generation
      description: |
        Returns every source the assembler would offer for this document type, so the RM can
        see what the system knows **before** it writes anything (FR-003) and deselect sources
        (FR-004). Deterministic — no model call, no vector search (R4).
      parameters:
        - { name: client_id, in: path, required: true, schema: { type: string, format: uuid } }
        - name: document_type
          in: query
          required: true
          schema: { $ref: '#/components/schemas/DocumentType' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssembledContext' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─────────────────────────── Sources ────────────────────────────
...
  /documents:
    post:
      tags: [Documents]
      summary: Generate a document draft
      description: |
        Runs the two-pass pipeline (R3): Grounding Pass builds the Evidence Ledger, Composition
        Pass produces schema-validated sections, then deterministic post-validation and Shariah
        screening run before anything is returned.

        **Fails closed.** If grounding fails, composition fails validation, a numeric literal
        cannot be traced to the ledger, or deterministic screening blocks, the response is an
        error and **no document is returned** (FR-037). A failure is still written to the audit
        trail (FR-039).

        The created document is always `status: DRAFT` and `shariah_status: PENDING_REVIEW`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, document_type]
              properties:
                client_id: { type: string, format: uuid }
                document_type: { $ref: '#/components/schemas/DocumentType' }
                meeting_notes:
                  type: string
                  description: Required for CALL_REPORT. Treated as untrusted data.
                source_document_ids:
                  type: array
                  items: { type: string, format: uuid }
                  description: Subset the RM chose to include (FR-004).
                client_record_ids:
                  type: array
                  items: { type: string, format: uuid }
                rm_instruction:
                  type: string
                  maxLength: 1000
                  description: |
                    Optional stylistic steer. Scoped to presentation only — it cannot
                    authorise a claim, alter screening, or affect approval (R7).
      responses:
        '201':
          description: Draft generated and screened
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '422':
          description: |
            Generation completed but failed deterministic validation — unresolvable evidence
            reference, untraceable numeric literal, or missing required section. No document
            is created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '451':
          description: |
            Shariah screening blocked the draft. The violation is reported; the draft is not
            returned (FR-016).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreeningBlockedError' }
        '503':
          description: Generation service unavailable. No partial content (FR-037).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Documents]
      summary: List documents visible to the caller
      parameters:
        - { name: client_id, in: query, schema: { type: string, format: uuid } }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DocumentStatus' }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/DocumentSummary' }
                  total: { type: integer }
  /documents/{document_id}:
    get:
      tags: [Documents]
      summary: Retrieve a document with its current version
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /documents/{document_id}/evidence/{claim_id}:
    get:
      tags: [Documents]
      summary: Inspect the evidence behind a cited claim
      description: |
        Returns the verbatim excerpt and its locator (page or character range) exactly as
        captured from the source. This is what the RM sees when inspecting a citation
        (FR-024). The excerpt is never paraphrased.
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: claim_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceClaim' }
        '404': { $ref: '#/components/responses/NotFound' }
  /documents/{document_id}/reject:
    post:
      tags: [Documents]
      summary: Reject the current draft
      parameters:
        - { name: document_id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: Rejected; recorded in the audit trail with actor and timestamp
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '409':
          description: Document is already APPROVED (terminal)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ─────────────────────────── Sections ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not permitted for this resource
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found, or not visible to this caller
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    DocumentType:
      type: string
      enum: [CALL_REPORT, CLIENT_PROFILE, CREDIT_MEMO_NARRATIVE, KYC_SUMMARY]
    DocumentStatus:
      type: string
      enum: [DRAFT, UNDER_REVIEW, REJECTED, APPROVED]
    ShariahStatus:
      type: string
      enum: [PENDING_REVIEW, CLEARED, FLAGGED]
      description: |
        Defaults to PENDING_REVIEW and is never set to CLEARED by the system. The system
        prepares documents for Shariah review; it does not clear them (Principle II).
    Confidence:
      type: string
      enum: [HIGH, MEDIUM, LOW]
    User:
      type: object
      required: [id, email, full_name, role]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        full_name: { type: string }
        role: { type: string, enum: [RM, TEAM_LEAD, COMPLIANCE, SHARIAH_REVIEWER] }
        team_id: { type: string, format: uuid, nullable: true }
    ClientSummary:
      type: object
      required: [id, client_reference, legal_name, sector, is_synthetic]
      properties:
        id: { type: string, format: uuid }
        client_reference: { type: string }
        legal_name: { type: string }
        trade_name: { type: string, nullable: true }
        sector: { type: string }
        relationship_since: { type: string, format: date, nullable: true }
        kyc_status: { type: string, enum: [COMPLETE, PENDING, EXPIRED] }
        is_synthetic:
          type: boolean
          enum: [true]
          description: Always true. Enforced by database CHECK constraint (FR-041).
    AssembledContext:
      type: object
      description: What the system knows, shown before it writes anything (FR-003).
      required: [client_id, document_type, sources]
      properties:
        client_id: { type: string, format: uuid }
        document_type: { $ref: '#/components/schemas/DocumentType' }
        required_inputs:
          type: array
          items: { type: string }
        sources:
          type: array
          items:
            type: object
            required: [source_id, source_type, label, is_external]
            properties:
              source_id: { type: string, format: uuid }
              source_type: { type: string, enum: [CLIENT_RECORD, UPLOADED_DOCUMENT] }
              source_system: { type: string }
              label: { type: string }
              effective_date: { type: string, format: date, nullable: true }
              is_external:
                type: boolean
                description: External-derived content is visually distinguished (FR-014).
        conflicts:
          type: array
          description: Surfaced, never silently resolved (FR-013).
          items:
            type: object
            properties:
              field: { type: string }
              values:
                type: array
                items:
                  type: object
                  properties:
                    source_id: { type: string, format: uuid }
                    value: { type: string }
...
    GapMarker:
      type: object
      description: |
        A first-class output state, not an error. Rendered visibly to the RM; blocks approval
        until resolved or explicitly acknowledged (FR-025).
      required: [field, label, resolved]
      properties:
        field: { type: string }
        label:
          type: string
          example: '[MISSING: audited turnover FY2025]'
        resolved: { type: boolean }
        resolution_note: { type: string, nullable: true }
    EvidenceClaim:
      type: object
      required: [claim_id, claim_text, source_type, verbatim_excerpt]
      properties:
        claim_id: { type: string }
        claim_text: { type: string }
        source_type:
          type: string
          enum: [CLIENT_RECORD, UPLOADED_DOCUMENT, MEETING_NOTES]
        source_id: { type: string, format: uuid, nullable: true }
        source_label: { type: string }
        locator:
          type: object
          description: Page range for PDFs, character range for text.
          properties:
            page_start: { type: integer, nullable: true }
            page_end: { type: integer, nullable: true }
            char_start: { type: integer, nullable: true }
            char_end: { type: integer, nullable: true }
        verbatim_excerpt:
          type: string
          description: The exact cited span. Never paraphrased.
        is_external: { type: boolean }
    DocumentSection:
      type: object
      required: [section_key, title, ordinal, evidence_refs, gaps, confidence]
      properties:
        section_key: { type: string }
        title: { type: string }
        ordinal: { type: integer }
        content:
          type: string
          nullable: true
          description: Null when the section is entirely a gap.
        evidence_refs:
          type: array
          items: { type: string }
          description: |
            claim_ids. Every one is validated against the ledger before the section is
            returned; an unresolvable reference converts the section to a gap (FR-011, R3).
        gaps:
          type: array
          items: { $ref: '#/components/schemas/GapMarker' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        is_rm_edited: { type: boolean }
        contains_external_data: { type: boolean }
    DocumentSummary:
      type: object
      required: [id, client_id, document_type, status, shariah_status, created_at]
      properties:
        id: { type: string, format: uuid }
        client_id: { type: string, format: uuid }
        client_reference: { type: string }
        document_type: { $ref: '#/components/schemas/DocumentType' }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        shariah_status: { $ref: '#/components/schemas/ShariahStatus' }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/DocumentSummary'
        - type: object
          required: [version_number, content_hash, sections, ai_generated, unresolved_gap_count]
          properties:
            version_number: { type: integer }
            content_hash:
              type: string
              description: Echo this to PATCH (If-Match) and to approve.
            ai_generated:
              type: boolean
              enum: [true]
              description: AI-generated content is labelled at all times (FR-020).
            template_version: { type: string }
            prompt_version: { type: string }
            model_id: { type: string, nullable: true }
            sections:
              type: array
              items: { $ref: '#/components/schemas/DocumentSection' }
            unresolved_gap_count:
              type: integer
              description: Approval is blocked while this is greater than zero (FR-025).
            screening:
              type: object
              properties:
                outcome: { type: string, enum: [PASS, BLOCKED, FLAGGED] }
                vocabulary_version: { type: string }
                findings:
                  type: array
                  items:
                    type: object
                    properties:
                      term: { type: string }
                      section_key: { type: string }
                      severity: { type: string }
                      rule_id: { type: string }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message:
          type: string
          description: |
            Plain, non-technical, and actionable — the RM must understand what happened and
            what to do next (FR-038, NFR-UX-02).
        detail: { type: object, additionalProperties: true }
    ScreeningBlockedError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            vocabulary_version: { type: string }
            findings:
              type: array
              items:
                type: object
                properties:
                  term: { type: string }
                  section_key: { type: string }
                  severity: { type: string }
                  rule_id: { type: string }
                  rationale: { type: string }
...